# Tune configs for NVFP4 CUTLASS dense GEMM